Validate boundaries for pgpGrab() when reading V4 signatures (#151)
In V4 signature packets there are multiple data sections, whose length
is stored as two bytes before the data. We've been checking the data
boundaries, but a truncated/malformed packet might be missing the
length data too. Add + use a little helper function to check the
length is actually there.