Signature check before header import on package reads
Until now we have been reading and importing the headers, and then
pull out the original immutable image from the header in order to
do signature checking on unknown packages. Those header manipulations
are relatively complex and thus vulnerable operatations that are,
to add to the insult, totally unnecessary as the header on disk is
just the way we want it for signature checking as it is.

This changes the package reading process rather fundamentally: we first
read (and sanity-check) the header blobs, then do signature checking on
the still un-imported blob, and only then try to import the headers
(which could still fail, at least in theory). This reduces the amount of
code exposed to unverified headers considerably: the deep header internals
are avoided entirely, and there's one less signature checking code-path
to worry about as it's now unified with the one used by headerCheck().

There's a subtle change in the "preferred" signature type: rpmpkgRead()
used to prefer DSA over RSA, headerSigVerify() prefers RSA. However
that's rather academic as both signature types never co-exist in
the same package.

For normal use, all this shouldn't change anything at all.