Handle buffer-too-small error from strftime()
Pass emsg around from the callers, actually emit an error if output
doesn't fit our buffer, allowing us to get rid of the old TODO note.

strftime() returning zero is not, generally speaking, always an error
because a format might be empty or the result could be empty, such
as %p/%P in many locales. However in our case the formats are known
to be non-empty.