Make signature range clearer and more flexible
Rename the siginfo payload field to range, which better explains
what it means. Also make it a bitfield which allows expressing
the actual combinations: modern signature/digest items are on
header only, rpm v3 ones are on header+payload and in the future
we might have payload-only digest(s) as well.

No functional changes intended.