Remove support for special legacy header handling from hdr.unload()
The refcounting and freeing seem rather suspicious, has probably
been broken for ages. As if anybody knew such an argument exists,
never mind what it might do or when it might be necessary. Should
this really be necessary, it would have to be automatic, and it'd
be almost 20 years too late for that...