Consolidate digest calculations to the read cycle
Process header and payload as separate entities when reading through
the newly created package for digest calculation. This consolidates
the digest calculations to one place and hopefully makes the overall
logic that little bit clearer too. In order to do this we need more
control than "Fread() to the end", add another helper to handle
jumping around + reading specified ranges.

Not supposed to affect functionality except for a slight change in
an error message nobody ever sees, this is still preliminaries for
things to come.