Implement payload digest (#163) verification as a part of signature
This is a bit wacky because the payload digest is not a part of the
signature header like everything else in "signature" checking, but
people are going to expect "rpm -K" to detect corrupted payload so
we better do it there anyway. We better do it during installation too
but that's another story.

It also takes a shortcut wrt the payload digest algorithm which is
supposed to be configurable and should be retrieved from
RPMTAG_PAYLOADDIGESTALGO, but as its currently hardwired to SHA256
on the build-side too it doesn't really matter ATM.