Make sure to reset file attributes for generated build-id directories.
[Note this patch is currently being tested in Fedora. See bug below.]

When creating the build-id directories we should reset the file attributes
to the defaults. Otherwise if the file list contained an %attr or %defattr
the directories would come out with the wrong mode.

Includes a testcase based on a spec by Igor Gnatenko that fails before
and Check that build-id directories are created with the right permissions
even if the spec file sets attrs explicitly.

https://bugzilla.redhat.com/show_bug.cgi?id=1432372

Signed-off-by: Mark Wielaard <mark@klomp.org>