Accept --nocaps also for the package installation
Makes it possible to work around trouble of installing packages that use
file capabilities in user namespaces.

It simply ignores the capabilities as opposed to setting a setuid bit.
This is a safer things to do, but possibly has a negative impact on some tools
(such as ping being usable only by the superuser).

https://bugzilla.redhat.com/show_bug.cgi?id=648654