Include system.h in ima and syslog plugins
Currently, there is no harm if config.h is not included in these files
because they are not sensitive to macros defined in config.h, but any
code added later or any plugin created using these plugins as examples
might be affected by these macros and therefore has to include config.h.

An example of bug when this header is not included properly can be seen
in the previous commit.

Signed-off-by: Gleb Fotengauer-Malinovskiy <glebfm@altlinux.org>