Eliminate rpmdsNotify() from the API + Python bindings
This is an internal helper for debug logging and should've never been
part of the API in the first place. We're removing all sorts of things
on this round anyway so it's a fine opportunity for removing this too.
It's not as if anybody used this outside rpm because it's ... not
exactly useful.