Add macro to force fsync() on close()
This patch adds a new macro that has rpm call fsync() before each close(). The
feature is off by default since most people will not want this, and is turned
on either with a macro or command-line flag.

There are a few reasons for wanting this:

1. To ensure that an RPM installation never fulls IO queues so much that
the production workload has to wait for it to be flushed when calling sync()
(think of databases)

2. For systems with very large numbers of disks to limit the amount of diskcache
an RPM installation can cause eviction of to limit IO thrash.

At Facebook, for the described workloads, we've found significant improvement on
our IO load when turning on this feature. Exact numbers are a bit hard to
quantify, but we are working on them. We see less threads of production
databases stacking up when yum/rpm/dnf are running, for example.

One word of caution: this will give good results on SSDs, but using this on
rotational disks won't make you happy.