In old packages, zero-length files have zeros for digest (RhBug:1352222)
Commit 8192746196745e15517e49230de183a0211017b6 enabled digest checks
for empty files, which is fine for modern era packages but rpm didn't
always generate a proper digest for them. Handle that as the special
case it is: preserve the digests checks as they are, but additionally
test and accept zeros too for MD5-era zero-length files.

(cherry picked from commit ab052ab775a789afc02cc9691b2e0526573bbea7)