Add some tests for digests, signatures and stuff
Add a pre-signed copy of the hello binary rpm, plus the public
and (passwordless) secret key for testing purposes. Run a couple
of simple rpmkeys --import and -Kv tests to see basics are working.
As a side-effect we now have a simple reproducability test as well.

It ain't much but it's more than we had before...