Implement basic signing tests for --addsign and --delsign
This is tricky and flimsy as we criss-cross over the fakechroot border:
getting gpg and gpg-agent to work in the fake root environment seems
non-trivial (mismatch about homedir expectations I think) so we
run signing outside the root and verification inside. And anything
on the outside needs to avoid absolute paths in results as those
would change from user to user etc. Still, better than not having it...