Avoid stale gpg-agents in test root
gpg2 wants to leave gpg-agent running in the background which is fine,
until we pull the rug from underneath it by nuking the test root it
was launched in, causing us to hang on --addsign tests. Shut down
any agents running in the test root before nuking it for an update.