Drop the ambiguous and bogus RSA/DSA algorithm info from sigcheck output
DSA is always SHA1 but PGP, which in rpm world means RSA, is certainly
not limited to MD5. To make matters worse, sha1 and md5 are ambiguous
as they can appear as standalone entities in this output. And to make
matters even worse, parenthesis are used to signal missing key as well,
but there are of course no missing keys for plain digests.