Add mode for updating with minimal writing (RhBug:783480)
How does the number of writes change?
I compared number of writes with and without minimal writing
for upgrade from RHEL 7.0 to RHEL 7.1, from RHEL 7.1 to RHEL 7.2, ...
In more details, the size of memory used for writing regular files.

The following table contains in the second column percentages of files
which can be only touched from all installed files. In the third column,
there are percentages of sum of their sizes.

                       |       Touched        |        Touched
                       |    Updated files     |     Updated bytes
------------------------------------------------------------------------
RHEL 7.0 -> RHEL 7.1   |         63 %         |          66 %
RHEL 7.1 -> RHEL 7.2   |         53 %         |          43 %
RHEL 7.2 -> RHEL 7.3   |         60 %         |          42 %
------------------------------------------------------------------------
F24 -> F25             |         63 %         |          40 %
F25-> Fraw             |         49 %         |          28 %

Does the speed change?
The update speed for classical disks or SSD almost does not change.

How it works?
If there is a file in the new package, which has the equal digest
as the corresponding file in the old package and the same file on
the disk (thus the contents are expected to be equal), rpm will not
install the whole file, but rather only upgrade the file's meta data.
In other cases, it will install the whole file.