Implement a callback hook into signature checking, use for rpm -K output
Add a new (internal for now) API rpmpkgVerifySignatures() for signature
checking, accepting an optional callback which will get called for
all the results we get. We only use it for rpmlog()'ing the results
in the ages old manner but could also be used to manipulate the results,
so its possible to implement external policy.