Move signature information into a table
Things get slightly worse until they get better: this causes an extra
lookup in the way things are right now, but this is more of an intermediate
state of things, we'll be switching the whole thing upside down in
the next steps.

Other notes:
All string data is now assumed hex and validated as such.
There are no *intentional* changes in the signature data, so any
incorrect type changes and the like would be copy-paste etc mistakes.
Test-suite is happy but that's not a guarantee.