Only look for known signatures and digests in signature checking
Take advantage of the newly added pieces to flip the verification
procedure upside down: we only look for what we can actually verify
and ignore everything else. Looping through the table also allows
us to skip disabled items without bothering to parse them, and
just makes the verification code that much nicer. As a side-note,
count for total number of failures instead of per verifyItems() call.

The bigger reason behind this change is to enable treating non-present
entries as verification failures where that makes sense, but that
behavior is not changed in this commit.