Look mom, no hands! Err, headers ...
Take advangate of hdrblobGet() to do rpm -K signature checking without
actually importing the headers, which is a relatively complicated
operation with all the region fiddling etc which isn't required
for signature checking at all. Another important point is that
both signature checking paths FINALLY operate on the same side of
header import which means it'll be possible to unify them to some
extent.