Assimilate rpmsinfoGet() into rpmsisetAppend()
rpmsinfoGet() wasn't such a good interface as it assumed 1:1 relation
between signature tags and number of signatures, this is something
we'll want to change. As there are no other users of rpmsinfoGet()
left, assimilate it into rpmsisetAppend() where arrays can be dealt
with. Provisions for arrays included but not enabled yet because
because there are no array types currently and append needs to handle
realloc to make arrays possible.