Minimal writing for %config files
How it works: if a file on disk has the same contents
as in the new package, rpm will only touch it
(originally rpm created it).

On the contrary to REG files, for detecting that both files
have the same contents, no additional comparison is needed.
Thus there is no reason why upgrade a package without
this modification. Hence this behavior is always used
(even without min_writes mode).