Implement SIG_DFL/SIG_IGN/SIG_ERR counterparts in rpmsq
RPMSQ_DFL is like passing NULL to rpmsqSetAction(), this already
set the default action but now we have a nice name for it too.
RPMSQ_IGN is the only special case needed in rpmsqSetAction() as
it corresponds to an actual function. Return RPMSQ_ERR for any
signal not supported by rpmsq.

This doesn't actually change any core behavior as such.