Use dribbles as heuristics for header/package output differentation
Installed packages will have dribbles so their index length is
larger than region length. The whole differentation is a bit arbitrary
since the stuff is always headers of course, but keeping in tradition...
This "fixes" the output for rpmReadPackageFile() to talk about
packages and not headers because it always deals with packages.