Use the signature checking code to read all package files
Both read package files and verify signatures, it only makes sense
to use the same code for both.

This does have significant consequences of course: we are now validating
ALL non-disabled header-only signatures and digests when reading
packages with rpmReadPackageFile() instead of just one that we figured
might be the best, if available. On one hand, it's a bit wasteful
since a good signature could be considered to make digests moot,
but then once we have real signature policies we might need to process
more than one signature to come to a conclusion anyway. For now,
just let it waste a few more cycles verifying redundancies.