Eliminate "tag" from rpmsinfo struct
Signatures are signatures and digests are digests no matter where
they come from, fade out the origins to the extent we can. The id
for hashable items is of course still derived from the tag and
can be backtraced (ditto for the disabler bits) but it's not
exactly full anonymity we need either, and eg id being related
to the tag could be useful for debugging.