Add RPMTAG_SOURCEPACKAGE to the source package header
We do that anyway when reading a header (see the applyRetrofits
function), so we might as well put it in right from the start.

Also, it's nice to have some indication for source packages in
the header instead of mis-using the absense of the SOURCERPM
element for that. We'll make use of that in the next commit.