Use a simple heuristic to test for source package headers
We do that only if neither RPMTAG_SOURCERPM nor RPMTAG_SOURCEPACKAGE
is set. Before, the code relied on checking the lead type, which
is somewhat ugly and error prone.

There are binary packages that are not generated by rpmbuild and do
not have a SOURCERPM tag set (because there is no source package),
there are also source packages that have a binary lead type. Both
might be seen as invalid packages, but that's life.

So we add a little heurisic: it is a source package if all the
files are in a single directory with an empty name.