Expand parameterized macro arguments one by one after splitting
This makes makes the behavior much more sensible eg in case an
argument expands to an empty strings, such as:
    $ rpm --define '%foo() 1:%1 2:%2' --eval '%foo %nil bar'
After commit 5adc56897b9da5dac49701f704ef54390db57c59 rpm was
evaluating that to "1:bar 2:%2" which is hardly what you'd expect.

Thanks to Michael Schroeder for pointing this out! (#217)