Simplify rpm -K non-verbose output
Spitting out misleading tag names encoded in various funny ways is not
that useful. Those who want details will run in verbose mode, in
non-verbose mode just note whether digests and signatures were present
and ok (or not) and the overall result. I'm not too happy about this
format either but at least it fades the tag details out of the picture.