Revert "Expand parameterized macro arguments one by one after splitting"
This essentially reverts commit 9ae7d1df313b7a2b9fd74fef5a176dcdce40b88b.
Leaving the testcase in place, just adjusting it to the expected
behavior where arguments that expand to empty are eaten away.

This actually matches what the shell does, so it follows the path of
least surprise in that sense - we emulate shell behavior with the
arguments etc already in many ways. What the shell has and rpm
has not is a way to pass those empty arguments by quoting. So
now we need a way to quote arguments - not just for passing empty
arguments but arguments containing whitespaces too.