Refactor an unnecessary intermediate helper out of signing code
Too many layers of helpers leftover from past times only getting
in the way of doing stuff. Also rename replaceSignature() arguments
to make it more obvious which is which.

This doesn't *really* change the signing, but it does optimize the
case where the package is already has a signature by the same key
and parameters: we generate the much cheaper V4 (ie header only)
signature and compare that, and only then do the V3 signature
which has to read the entire payload. This can make a huge difference
on large packages.