Use C.UTF-8 locale, if available
Since some distributions ship a glibc with support for the C.UTF-8
locale.  This commit adds a check in configure for it and sets it as
a default if it is available on the system.

There are more and more applications coming up, which imply support
for UTF-8 encoding present in the runtime enviroment, during tests for
instance.  So rpm should provide support for such cases out of the box.
And since the new C.UTF-8 locale is fully backwards compatible with the
tradional ASCII-based C locale, there are no incompatibilities to expect.

If there is no support for this new locale, e.g. not using a glibc version
having it present, we keep the old C locale by default.