Bury get_fskpass() inside rpmsign utility
librpm is not in the business of providing terminal utility functions,
file signing might well need to ask for passwords but it doesn't
have to be a non-prefixed function in a shared library. The library
provides means to *pass* the password and its up to calling applications
to ask for it if needed.