Conditional compilation sanity for file signing
Give a nice and clear error message if file signing is requested
on a build that doesn't support it, and ifdef the entire
contents of includeFileSignatures() so we wont try to compile
in a call to rpmSignFiles(), which doesn't get built at all
due to makefile conditionals when imaevm is not enabled.
On a related note, drop the unnecessary conditional and error message
from rpmsignfiles.c: the whole file does not get built if imaevm
is not enabled so there's no need for these.