don't use undefined $NAME, but rather use array member of $NAMES
This fixes problem with not being able to pass indidividual names for
the .lang file to include. @ffesti pointed out in #12 that ${NAME}
would not be available at end of script.
This is false as variables stays resident to whatever last defined
as, so ${NAME} won't go undefined as if it were to go out of bounds..