Add support for new package declarations in Perl 5.12+
This change allows the perl.prov script to detect new variants of Perl
package declarations that have been added in recent years.

In Perl 5.12 we got version numbers as part of the package declaration.

  package Foo 1.0;

  package Bar v1.0.0;

And in Perl 5.14 we got the package block syntax.

  package Foo {...}

  package Bar 1.0 {...}

  package Baz v1.0.0 {...}

We noticed this problem recently because Dist::Zilla, a popular Perl
module, started using a version number as part of the package
declaration and perl(Dist::Zilla) could no longer be resolved.