Avoid double macro definition on first --define
Commit 71327238150d689f0b4cfe7be8cd8addb0ccc78a introduced --predefine
strangely mixed up with -D. Part of the mixup was later eliminated
in commit 3d657d5cc22652c71bc15a56b6b4462e228526d6 but what was left
behind is inserting a predefine on first regular --define, which
among other things causes that macro to be doubly defined, which
in turn causes funny side-effects when you try to --undefine that macro.

Also add a test-case for the behavior, prior to this change the --undefine
in the first command appeared to have no effect.

Thanks to Pavlina Moravcova Varekova for spotting and reporting this.