Rip out ndb's check if the database file was replaced
The implementation did not work correctly when a chroot was done.
As the code didn't deal with the index database at all and was
thus not really useful, it's better to remove it instead of
trying to fix it. Other databases like BerkeleyDB also do not
check for this condition.