Change hdrNumCmp to use the tagNum as secondary sort key
This brings rpmdbSortIterator in line with its documentation.

Note that there is a catch: this change also makes the dbiIndexSetPrune
function work different! It used to delete all items that had a
matching hdrNum, after the change only the one is deleted that also
has a matching tagNum. I consider this a bug fix.
Note that dbiIndexSetPrune is only used in the db3.c backend to remove
index entires, if we really want the old behavior we should add a
new dbiIndexSetPruneHdrnum function.

Also remove assertion when trying to sort an empty set.