Shift macro scoping level values as they are defined
Some of scoping level values are defined in rpmmacros.h:
 #define RMIL_DEFAULT    -15
 #define RMIL_MACROFILES -13
...
 #define RMIL_GLOBAL     0

Before this patch RPM set these values decreased of 1 (for all macros).
Because of the (level - 1) in
    pushMacro(dst, me->name, me->opts, me->body, (level - 1), me->flags);
in function copyMacros scoping level for some CMDLINE macros was decreased of 1.
Because of the (level - 1) in
    pushMacro(mb->mc, n, o, b, (level - 1), ME_NONE);
in function doDefine scoping level values of all other macro types and
some CMDLINE macros was decreased of 1.
The patch shifts scoping level values back to the numbers defined
in rpmmacros.h.

This was spotted when investigating a regression introduced in
commit 237f35f16f978bb20e4cafd78abfab1f2e5db480 where locally scoped
scoped macros did not get deleted on scope closing.
This also fixes the regression in question.

The first test in this patch tests automatic deleting of %defined macros.
The second test checks whether the behavior changed by
237f35f16f978bb20e4cafd78abfab1f2e5db480 is still working correctly.

Thanks to Panu Matilainen for investigation and proposal of the main part
of the patch!