Really ignore unknown signature tags (RhBug:1480407)
Rather embarrassingly, rpm 4.13 fails rpmkeys -K signature verification
of rpm 4.14 packages because it doesn't really ignore unknown
signature tags like it claims to, and should. And so it reports a
bogus failure on the new SHA256 header-only digest. This bug as it is
only exists in 4.13.x, older versions are fine and >= 4.14 only
ever looks for tags it knows about.