Soften the 4.4.x COMPAT error into a warning for 4.14.x
People have already been updating their code to build with 4.14.x API
with the alpha that did not have this check, suddenly erroring out if
if a no-op define is left in place seems like a bit severe here and
just make people (even more) grumpy for no good reason.

Just warn on 4.14.x, the next version will have it as an actual error.