Add support for 'unless' rich dependencies
An (A unless B) dependency implements (A and not(B)). This kind is useful
for "or" type dependencies, e.g. "Conflicts" or "Supplements".

As "Conflicts: (A unless B)" is equivalent to "Requires: (B if A)", I
thought this type is not needed. But there is no such equivalence
for Supplements, thus the change in mind.

Like with "if" we also have a syntactic sugar "else" flavor:
(A unless B else C) is the same as ((A unless B) or (B and C))

This commit also makes the "else" handling code in depends.c much
easier to understand.

(cherry picked from commit 37982efbacdc2ccc00cea988a049dfa32749936f)