Always execute scriptlet callbacks with owning header (RhBug:1485389)
Triggers and file triggers can and do execute scriptlets from installed
packages which are not part of the current transaction and thus have no
associated transaction element, making the scriptlet callbacks
inconsistent and cumbersome for API users.

Create a fake transaction element for the poor orphan scriptlets lacking
one to make it consistent (of course, creating rpmte's with all their
associated data just to pass a header pointer along is ridiculously
expensive but *shrug*).

Regular triggers used to execute in the context of the triggering
transaction element, make them run with the owning trigger header too.