Enable python build during dist-check
The test-suite fails if python is not enabled. An alternative solution
could be disabling python tests when not enabled, but the python
tests cover things that are not covered elsewhere so especially
for dist cutting the tests are quite important.

(cherry picked from commit 5763f40c2130936f398d561e7ea6de66fb9da086)