Always calculate disk space info, only don't report them if filtered
In the distant past when rpm just stat()'ed all mounted filesystems,
it was quite necessary to have the option of disabling the fs stats.
However now that we only stat() filesystems we actually use, the
situation is quite the reverse: if stat() eg hangs, we wouldn't be
able to complete the transaction anyway so it's better to find that
out before the transaction even starts, regardless of problem filters.

Also the filesystem info sets could be used to track other things besides
space, such as suid/sgid availability, selinux and file caps support etc.