Fix file triggers failing to match on some packages (MgBug:18797)
Directory names as stored in RPMTAG_DIRNAMES are not sorted when
separated from basenames, so binary search has no chance of working.
While a linear search on the dirnames would be guaranteed to find *some*
matches when they exist, it could still miss some results as the
matches are not guaranteed to exist in a neat low-high range.

Construct the entire pathname for prefix comparisons to ensure sorted
paths and adjust the file trigger testcase to cover this too.

Thanks to Pascal Terjan for initial investigations and Thierry
Vignaud for providing a workable reproducer.